clean up circular references in ts
